requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-tables2==2.5.3
django-filter==23.3
django-chartjs==2.3.0
